RD MB

+++ Assets/chadiik/topics/reactionDiffusion/ReactionDiffusion.shader
@@ -16,68 +16,68 @@
 
     #include "UnityCustomRenderTexture.cginc"
 
-	float4 _Convolution;
-	float _A;
-	float _B;
-	float _Feed;
-	float _Kill;
-	float _DT;
+	half4 _Convolution;
+	half _A;
+	half _B;
+	half _Feed;
+	half _Kill;
+	half _DT;
 
-	float4 frag(v2f_customrendertexture i) : SV_Target
+	half4 frag(v2f_customrendertexture i) : SV_Target
 	{
-		float2 uv = i.globalTexcoord;
+		half2 uv = i.globalTexcoord;
 
 		// pixel size
-		float2 texel = float2(1. / _CustomRenderTextureWidth, 1. / _CustomRenderTextureHeight);
+		half2 texel = half2(1. / _CustomRenderTextureWidth, 1. / _CustomRenderTextureHeight);
 
-		float4 cell = tex2D(_SelfTexture2D, uv);
+		half4 cell = tex2D(_SelfTexture2D, uv);
 
 		// 3x3 convolution:
-		float cellWeight = _Convolution.x;
-		float adjacentWeight = _Convolution.y;
-		float diagonalWeight = _Convolution.z;
+		half cellWeight = _Convolution.x;
+		half adjacentWeight = _Convolution.y;
+		half diagonalWeight = _Convolution.z;
 		
 		// can be fixed2
 		// adjacent cells
-		float4 left = tex2D(_SelfTexture2D, uv - float2(texel.x, 0.));
-		float4 right = tex2D(_SelfTexture2D, uv + float2(texel.x, 0.));
-		float4 up = tex2D(_SelfTexture2D, uv - float2(0., texel.y));
-		float4 down = tex2D(_SelfTexture2D, uv + float2(0., texel.y));
+		half4 left = tex2D(_SelfTexture2D, uv - half2(texel.x, 0.));
+		half4 right = tex2D(_SelfTexture2D, uv + half2(texel.x, 0.));
+		half4 up = tex2D(_SelfTexture2D, uv - half2(0., texel.y));
+		half4 down = tex2D(_SelfTexture2D, uv + half2(0., texel.y));
 		// diagonal cells
-		float4 leftUp = tex2D(_SelfTexture2D, uv - texel);
-		float4 rightDown = tex2D(_SelfTexture2D, uv + texel);
-		float4 rightUp = tex2D(_SelfTexture2D, uv + float2(texel.x, -texel.y));
-		float4 leftDown = tex2D(_SelfTexture2D, uv + float2(-texel.x, texel.y));
+		half4 leftUp = tex2D(_SelfTexture2D, uv - texel);
+		half4 rightDown = tex2D(_SelfTexture2D, uv + texel);
+		half4 rightUp = tex2D(_SelfTexture2D, uv + half2(texel.x, -texel.y));
+		half4 leftDown = tex2D(_SelfTexture2D, uv + half2(-texel.x, texel.y));
 
 		// convolution result
-		float4 convolution = 
+		half4 convolution = 
 			cell * cellWeight + 
 			(left + right + up + down) * adjacentWeight + 
 			(leftUp + leftDown + rightUp + rightDown) * diagonalWeight;
 
 		// http://www.karlsims.com/rd.html
 		// Reaction (A) in red/x
-		float a = cell.x;
+		half a = cell.x;
 		// Reaction (B) in green/y
-		float b = cell.y;
+		half b = cell.y;
 
-		float abb = a * b * b;
+		half abb = a * b * b;
 		
-		float rdA =
+		half rdA =
 			a +
 			( 
 				_A * convolution.x - abb +
 				_Feed * (1. - a)
 			) * _DT;
 
-		float rdB =
+		half rdB =
 			b +
 			(
 				_B * convolution.y + abb -
 				(_Kill + _Feed) * b
 			) * _DT;
 
-		float4 color = float4(saturate(rdA), saturate(rdB), 0., 0.);
+		half4 color = half4(saturate(rdA), saturate(rdB), 0., 0.);
 
 		return color;
 	}
